+++ src/main/java/carpet/CarpetServer.java
@@ -38,7 +38,7 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.commands.PerfCommand;
 import net.minecraft.server.level.ServerPlayer;
-import org.thinkingstudio.bedsheet.util.NeoHooks;
+import org.thinkingstudio.bedsheet.loader.FoxifiedLoader;
 
 @SuppressWarnings("removal")
 public class CarpetServer // static for now - easier to handle all around the code, its one anyways
@@ -141,7 +141,7 @@ public class CarpetServer // static for now - easier to handle all around the co
         if (environment != Commands.CommandSelection.DEDICATED)
             PerfCommand.register(dispatcher);
         
-        if (NeoHooks.isDevelopmentEnvironment())
+        if (FoxifiedLoader.isDevelopmentEnvironment())
             TestCommand.register(dispatcher);
         // todo 1.16 - re-registerer apps if that's a reload operation.
     }