+++ CHANGELOG.md
@@ -1,4 +1,2 @@
 ## Changelog
-- port to 1.21.3
-- rework `customSticky` mixin
-- fix conflict with worldedit
+- change: `NeoHooks` -> `FoxifiedLoader`